implemented the compareTo method

+++ src/Job.java
@@ -1,19 +1,19 @@
 public class Job{
     int id; //id of the Job
-    int time; //time that takes for the job to be completed
+    double time; //time that takes for the job to be completed
 
-    Job(int Job_id,int time){
+    Job(int Job_id,double time){
         this.id=id;
         this.time=time;
     }//constructor
 
     void setIdd(int id){this.id=id;} //setting the value of the id
 
-    void setTime(int time){this.time=time;}//setting the value of the time
+    void setTime(double time){this.time=time;}//setting the value of the time
 
     int getId(){return id;} //returning the value of the id
 
-    int getTime(){return time;}//returning the value of the time
+    double getTime(){return time;}//returning the value of the time
 
     String toString(){
         return "Id: "  + id.toString() + " time: " + time.toString();
